fix two things:
- completely broken determination of who goes next
- handling of dead and incapacitated, which previously just reverted
  the game to the previous creature when the next creature was dead/incap.

however, the fix for dead/incap really shows that the way I'm handling
it is completely wrong -- we *should* put the game into a state where the dead
or incap player is "current", but disallow them from doing anything. This
is so we don't go into an infinite loop when all players are dead!