Refactor to pass in origin-creature wherever effects are applied.
Notably, this means I added a CreatureName field to AppliedCondition
so that when a condition applies an effect, it remembers who applied
the condition. This means more Maybe-ing since we have to look up
origin creature name in the game in more places. (including the
Creature itself in the AppliedCondition would be wrong because then
effects relying on the origin creature would be "snapshotted" to
the original state of the creature instead of the updated state).