fix GameLogs for CreateCreature in UI
I forgot that the GameLog still serializes a Creature (not a DynamicCreature) on the server.
So I made a new type, CreatureData, on the client, which it uses to parse that structure.
The main difference is that the dynimac stuff like can_act and can_move isn't on it. For now I just parse out bare-bones stuff.