allow providing RPI URL as an Elm "Program Flag"
Sadly, Program Flags can't be specified when using either elm make or elm reactor -- the only way to provide them is with a JavaScript call which we embed in a custom HTML file.
So now, we have HTML files in our PTUI. :-(

So, there's also now a build tool that runs elm make to generate .js files, and also spits out Player.html and GM.html files with URL provided on the command line.