Implement versioned polling
This makes the polling MUCH more reliable. Yay!

Polling now requires the client to pass in a "version" of the app that it has.

This version number is specified as the length of the snapshots and the length of the logs in the latest snapshot.

This is compared to the current state on the server. If there's a match, we do the normal polling. If there's a mismatch, we immediately return the current application.

This avoids race conditions where the client would try to poll just after some change was made, and miss out on the update.