make the attr check UI a lot nicer
Now when you trigger an attribute check, the results show up in a modal dialog box.

This is based on the Game Log system; the results are really just the rendered game logs from that specific command.

I also simplified the result of the RPI server to only return GameLogs, not (Game, GameLogs).