A FULLY WORKING "Give" UI in TS/React!
This sends the GameCommand via the Elm app, but it should be easy to eventually switch that over to doing it straight from TypeScript code.

I've also added a new "PTUI" class which keeps both the `PTTypes.App` and the Elm App. This allows us to pass a single object down our React Component hierarchy that can both read all the game state and send commands.