no more use of indexing operation for objects or arrays, because it's not typesafe
I discovered that typescript allows foo[bar] to return the value of the `foo` object without unioning it with `undefined`. This is horrible! It leads to all sorts of runtime errors that aren't caught by the compiler.

I added `get` and `idx` functions which replace `obj[key]` and `array[idx]` and return T | undefined.