render a menu when clicking on creatures that have `actions` defined in the Grid code.
This is a basic proof of concept of rendering a menu. There are two things I want to do from here:

First, I want to actually implement the "Move" action (for moving a creature we've clicked on)

I also want to refactor the code so it uses Redux instead of manual propagation of `onClick` handlers through multiple layers of React components.
This is the kind of situation where Redux/TEA are nice: communication between relatively disparate parts of the UI -- and avoiding the use of closures as props means that our components will be more efficient, because you can't compare closures for equality.
Right now we are constantly rerendering the entire Grid every time any of the props change (e.g. clicking on a creature to bring up a menu), becaues that change is causing an (equivalent) closure to be regenerated and passed into the child component's props.