a few redux ergonomics things
- improve the type of the `dispatch` function in connectRedux so it actually requires our Action type
- make connectRedux work with functions as well as classes

Also tweaked the rendering of the creature dropdown menu, showing the creature name and a "Close" item regardless of whether there are any actions enabled for that Creature. This lets you see what a creature's name is!

and I set up a pipeline for the RequestMove action, unimplemented as yet.