Implemented narrow/wide view in ReactPlayer
This puts the map in a tab when the screen is too narrow.

And there were NO bugs with the SVG PanZoom integration, and no extra work needed, because the react component lifecycle is actually useful, unlike Elm!

TODO: scale the narrow view so it fills the width of the device.