fix narrow view... again
svg-pan-zoom does not work if the svg is display:none, because it needs to do size calculations. so instead we special-case the Map tab to *always* render, but with visibility: hidden. This required tweaking a bunch of css again, now using relative/absolute positioning for the tab contents.

But it seems to be working... on desktop browsers. Yet to test mobile.