feature: creature selection for scenes
this involved a big change to the campaign tree code, just as it did in the original elm code, though I think it turned out nicer.

The one sad thing is that I decided to get rid of the inline creature card in the campaign tree just because it was too much of a pain to fit it into my more generalized design for the tree. Might bring it back later, but I don't think it was *particularly* elegant or anything...