add background_image_url to *Map*
This is to support actual map images, as opposed to Scene backgrounds.

Scene backgrounds are meant to be abstract or evocative imagery that goes along with the scene (though I'm not sure how useful it will be in the long run) -- I imagine it mostly being used for mapless scenes, though a "Handout" system may be better for that.

Map backgrounds are meant to be actual terrain maps in image form, which are then augmented by the P&T Map.terrain data to indicate the actual paths available within that tactical map. In the UI, they will be scaled/panned along with the terrain.

TBD: UI support. This shouldn't break the RPI since I am defaulting the background image to "".