just some tinkering about how to manage World
since World is not Send, there is no way to use it in multiple threads, even with Mutex. This is because it uses internal mutability in a thread-unsafe way (i.e. it uses Rc).

So I think the only thing I can do is to manage all access to the World in a single thread, and deal with Rocket's multi-threaded handlers by passing messages to that single worker thread.