Various updates to frontend dependencies, and trying to switch to vite.
Unfortunately some of the old dependencies I used are long-dead now, and their functionality has been hacked out but needs to be replaced:

- WindowSizeListener
- SplitPane

I also need to get GM vs Player UIs working again, but before that I need to get *anything* rendering.
The frontend will fetch the initial state of the game but it's currently failing to poll and not showing anything.