Remove all filesystem interaction and JSON/YAML parsing/dumping from the core pandt game library.
Now all file I/O and serializing and deserializing happens at the PTRPI level. This will allow me to more easily try other storage backends like Google Firestore or whatever.

The only GameCommand that had to change was LoadModule, which previously took a file name and loaded it from disk. Now there is a specific web endpoint for loading a module at a path, and the PTRPI does the file I/O and just passes the entire deserialized Game module to the LoadModule GameCommand

This also simplifies stuff like perform_command and gets rid of a bunch of error variants from GameError