proof of concept for the client getting a game object from the new endpoint.
I think I actually don't want to use Cookies for pt-id-token any more? at least not for the PTRPI - maybe it makes sense to use them just in the PTUI javascript as a way to cache the token. I'm not sure if the way I'm setting them would even work in actual deployment.