Don't try to do authentication while creating WebSockets
WebSocket creation requests can't have any extra headers passed in (this
is a limitation of the web browser WS API). So the client now sends a
request to /request-websocket/{game_id} (which is authenticated) to get
a temporary token, and then hits /ws/{token} to connect to the Durable
Object with a WebSocket.