Fix the definition of Create/Edit Class/Ability to not do random side-effects during log application.
I messed up in modeling the logs for Create/Edit Class & Ability.

Application of GameLogs must never have any random ID generation in them (or any other side-effects). This is because we need to replay these logs from a snapshot and get a deterministic state.

This implies that GameLogs must have everything they need to modify a game object deterministically -- e.g., GameLog::CreateClass can't take a ClassCreation, it has to take a full Class that includes the ID!

At the same time, I'm changing EditClass and EditAbility to wrap just a Class or an Ability, instead of an ID & ClassCreation or ID & AbilityCreation. This is mostly just a bikeshedding thing; Class already *is* an ID + ClassCreation.

I got lucky here because I didn't happen to test out class creation or editing in production, but I know that very soon I will need to implement a migration system for when I need to make these kinds of changes after there are already GameLogs in the wild.