Try using the dioxus component system
Not sure I'm a fan of this - I really don't see the point of copying the Rust code for the button
component into my source code. This seems like it blindly copies the way some Javascript component
libraries work without considering that there are better ways to do it in Rust. (like, a crate with
features for each component...)